Added missing methods to AdapterInterface
Added a BaseAdapter to avoid code duplication in future adapters
