+++ src/Phinx/Db/Adapter/AdapterInterface.php
@@ -524,6 +524,24 @@ interface AdapterInterface
     public function dropDatabase($name);
 
     /**
+     * Creates the specified schema or throws an exception
+     * if there is no support for it.
+     *
+     * @param  string $schemaName Schema Name
+     * @return void
+     */
+    public function createSchema($schemaName = 'public');
+
+    /**
+     * Drops the specified schema table  or throws an exception
+     * if there is no support for it.
+     *
+     * @param string $schemaName Schema name
+     * @return void
+     */
+    public function dropSchema($schemaName);
+
+    /**
      * Cast a value to a boolean appropriate for the adapter.
      *
      * @param mixed $value The value to be cast
